+++ modules/properties/core/src/sap/commerce/toolset/properties/custom/settings/state/CxCustomPropertyTemplatesState.kt
@@ -16,10 +16,11 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.actionSystem
+package sap.commerce.toolset.properties.custom.settings.state
 
-object HybrisActionPlaces {
-    @Deprecated("review this usage, migrate to LoggersConstants")
-    const val LOGGERS_TOOLBAR = "SAP.Loggers.View"
-    const val PROPERTIES_TOOLBAR = "SAP.Properties.View"
-}
+import com.intellij.util.xmlb.annotations.Tag
+
+@Tag("HybrisCustomPropertyTemplates")
+data class CxCustomPropertyTemplatesState(
+    @JvmField val templates: List<CxCustomPropertyTemplateState> = emptyList(),
+)